--- a/Common/3dParty/openssl/fetch.sh
+++ b/Common/3dParty/openssl/fetch.sh
@@ -4,5 +4,5 @@ SCRIPTPATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd "$SCRIPTPATH"
 
 if [ ! -d openssl ]; then
-  git clone https://github.com/openssl/openssl.git
+  git clone --depth=1 https://github.com/openssl/openssl.git
 fi